Move parsing logic
