+++ grazer/config.py
@@ -7,6 +7,8 @@ import logging
 from functools import reduce
 from jinja2 import Template
 
+from grazer.core.parsing import create_node, parse
+
 
 logger = logging.getLogger(__name__)
 
@@ -82,46 +84,11 @@ class Mapping(object):
 
     def __init__(self, key, pattern):
         self.key = key
-        self.path = [self.create_node(part)
+        self.path = [create_node(part)
                      for part in pattern.split("/")]
 
-    def create_node(self, data):
-        tag_part = r"(?P<tag>\w+)"
-        attr_part = r"(?P<q>\[(?P<attr>\w+)=(\"|\')(?P<val>.+?)(\"|\')\])?"
-        selector_part = r"(\{(?P<selector>\d+)\})?"
-        p = tag_part + attr_part + selector_part
-        patt = re.compile(p)
-
-        m = patt.match(data)
-        tag = m.group("tag")
-
-        if m.group("q"):
-            q = {m.group("attr"): m.group("val")}
-        else:
-            q = None
-
-        def selector(lst):
-            s = m.group("selector")
-            if s:
-                sel = int(s)
-                return [lst[sel]] if sel < len(lst) else []
-            else:
-                return lst
-
-        def node(root):
-            return selector(root.findAll(tag, q))
-
-        return node
-
     def parse(self, root):
-        context = [root]
-        for path in self.path:
-            node = context.pop()
-            logger.debug("Using path: {0} entering context {1}".format(path, node))
-            for out in path(node):
-                context.append(out)
-        return [(self.key, result.text, result.attrs)
-                for result in context]
+        return parse(self.key, self.path, root)
 
 
 class Config(object):
